Split dropzone thumbnail out
