+++ src/components/pages/student-submit/form/image-dropzone.js
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import { css } from '@emotion/core';
 import styled from '@emotion/styled';
-import Icon from '@leafygreen-ui/icon';
 import { useDropzone } from 'react-dropzone';
-import Button from '~components/dev-hub/button';
 import { H5, P, P3 } from '~components/dev-hub/text';
 import { size } from '~components/dev-hub/theme';
+import DropzoneThumbnail, { THUMBNAIL_WIDTH } from './dropzone-thumbnail';
+
+const DROPZONE_HEIGHT = '232px';
 
 const greyText = theme => css`
     color: ${theme.colorMap.greyLightTwo};
@@ -27,14 +28,6 @@ const GreyP3 = styled(P3)`
     ${({ theme }) => greyText(theme)};
 `;
 
-const RemoveButton = styled(Button)`
-    position: absolute;
-    top: 0;
-    right: 0;
-    padding: 0;
-    color: black;
-`;
-
 const Dropzone = styled('div')`
     align-items: center;
     background: ${({ theme }) => theme.colorMap.greyDarkThree};
@@ -43,42 +36,15 @@ const Dropzone = styled('div')`
     cursor: pointer;
     display: flex;
     flex-direction: column;
-    height: 232px;
+    height: ${DROPZONE_HEIGHT};
     justify-content: center;
     text-align: center;
-`;
-
-const ThumbnailWrapper = styled('div')`
-    border-radius: 10px;
-    width: 96px;
-    height: ${size.xlarge};
-    padding: 4px;
-    margin: 0 ${size.mediumLarge} ${size.xsmall} 0;
-    grid-row-start: 2;
-    border: 1px dashed ${({ theme }) => theme.colorMap.greyLightThree};
-    :last-of-type {
-        margin-right: 0;
-    }
-`;
-
-const ThumbnailContent = styled('div')`
-    display: flex;
-    height: 100%;
-    min-width: 0;
-    overflow: hidden;
-    position: relative;
-`;
-
-const Image = styled('img')`
-    display: block;
-    height: 100%;
-    width: auto;
 `;
 
 const ThumbnailGrid = styled('div')`
     display: grid;
     grid-template-rows: ${size.medium} ${size.xlarge};
-    grid-template-columns: repeat(6, 96px);
+    grid-template-columns: repeat(6, ${THUMBNAIL_WIDTH});
     text-align: center;
     row-gap: 4px;
     column-gap: ${size.mediumLarge};
@@ -88,29 +54,6 @@ const ThumbnailGrid = styled('div')`
 const ImageLabelText = styled(P3)`
     color: ${({ theme }) => theme.colorMap.greyLightTwo};
 `;
-
-const Thumbnail = ({ file, removeImage }) => {
-    const isFile = !!file;
-    const preview = isFile && file.preview;
-    return (
-        <ThumbnailWrapper>
-            <ThumbnailContent>
-                {preview ? (
-                    <div>
-                        <Image src={preview} />
-                        <RemoveButton onClick={removeImage}>
-                            <Icon
-                                glyph="XWithCircle"
-                                size="small"
-                                fill="#D8D8D8"
-                            />
-                        </RemoveButton>
-                    </div>
-                ) : null}
-            </ThumbnailContent>
-        </ThumbnailWrapper>
-    );
-};
 
 // Adopted from https://react-dropzone.js.org/#section-previews
 const ImageDropzone = ({ maxFiles = 6 }) => {
@@ -131,7 +74,7 @@ const ImageDropzone = ({ maxFiles = 6 }) => {
     });
 
     const thumbs = files.map((file, index) => (
-        <Thumbnail
+        <DropzoneThumbnail
             file={file}
             key={file ? file.name : index}
             removeImage={() => {
